Automatic login using WebEngine works
This commit allows us to automatically login via the webengine.


NOTE: We are still facing and issue, where the returned HTML is incomplete, this is because we have yet to figure out a way to confirm that the webpage has fully loaded.

Side Effects: There is a dedicated class that is being used to run this, this class has been omitted. 

Authors: Philipp