ParseSpecificTag() returns a JSON object
This commit makes the parseSpecificTag() return a JSON object, this allows us to format the tags.

Side effects: This may cause some merge conflicts because JUnit tests and the Interface changes the parseSpecificTag() return type.

NOTE: The next step is to allow the developer to create a file with tags that they want to parse.

Authors: Philipp, Jay