Partial fix to thread issue
This commit will introduce a partial fix for the threading issue.

NOTE: There is significant slowdown with this fix, we are not sure why this us happening at the moment, we will investigate this further once the connection has been reworked.

Side effects: There may be some issue when pulling

Authors: Philipp, Jay